Adds basic documentation and Makefile to build PDFs
Adds an extra folder called ''Documentation'' that contains a Makefile to generate the documentation. The documentation's source is in the src/ folder.

The documentation isn't written, yet, so to show how it will look like, there is a ''Lorem Ipsum'' intro with some text. The png file is stored is needed for the documentation PDFs.