Recognizes EOF now
The problam was, that we didn't check how many bytes we actually read. The read function returns the amount of bytes it has read. If this number is lower than the chunk of bytes that one wanted to read in, this means that the file's end has been reached.
So in this case a counter is used to only traverse the bytes from the file and then return and eof.

Closes #6, closes #1